# pkg/apigateway/spec/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "V2.Enjoyfood Project",
    "version": "0.0.1",
    "contact": {
      "name": "Legends of three kingdoms Project",
      "url": "https://ltk.ranklier.tech"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/oauth2/token": {
      "post": {
        "summary": "快速登录",
        "operationId": "SignInOrSignUp",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ltkSignInOrSignUpResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ltkSignInOrSignUpRequest"
            }
          }
        ],
        "tags": [
          "TokenService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "ltkSignInOrSignUpRequest": {
      "type": "object",
      "properties": {
        "nick_name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "ltkSignInOrSignUpResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "The Authorization header must be set to Bearer followed by a space and a token. For example, 'Bearer vHUabiBEIKi8n1RdvWOjGFulGSM6zunb'.",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}